Revert "branch-A-MoreOOP commit"
This reverts commit 5a8abc6257f860fe79197473b49a4afb2a76093f.
