+++ src/main/java/Task.java
@@ -1,4 +1,3 @@
-
 public class Task {
     private String text;
     private String type = "";
