+++ src/main/java/TaskArray.java
@@ -1,4 +1,3 @@
-
 import java.util.ArrayList;
 
 public class TaskArray {
